+++ QuizzBankBE.DataAccessLayer/DataObject/QuizBank.cs
@@ -34,7 +34,7 @@ public partial class QuizBank : IAuditedEntityBase
 
     public int? AuthorId { get; set; }
 
-    public int? DefaultMark { get; set; }
+    public float? DefaultMark { get; set; }
 
     public virtual User? Author { get; set; }
